Moved complexity to a separate class
This way the codebase should be easier to understand and work with
