+++ subprojects/shipkit/src/main/groovy/org/shipkit/internal/gradle/git/GitAuthPlugin.java
@@ -8,9 +8,8 @@ import org.shipkit.gradle.configuration.ShipkitConfiguration;
 import org.shipkit.internal.gradle.configuration.ShipkitConfigurationPlugin;
 import org.shipkit.internal.gradle.git.tasks.IdentifyGitOriginRepoTask;
 import org.shipkit.internal.gradle.util.TaskMaker;
-import org.shipkit.internal.util.ExposedForTesting;
 
-import java.text.MessageFormat;
+import static org.shipkit.internal.gradle.git.GitHubUrlBuilder.getGitHubUrl;
 
 /**
  * This plugin is used for internal purposes, it does not add any user-visible, public behavior.
@@ -51,24 +50,6 @@ public class GitAuthPlugin implements Plugin<Project> {
                 action.execute(new GitAuth(repoUrl, writeToken, identifyTask.getOriginRepo()));
             }
         });
-    }
-
-    //TODO SF move getGitHubUrl & _getGitHubUrl to separate class, along with unit tests
-    static String getGitHubUrl(String ghRepo, ShipkitConfiguration conf) {
-        return _getGitHubUrl(conf.getGitHub().getWriteAuthUser(), ghRepo, conf.getLenient().getGitHub().getWriteAuthToken());
-    }
-
-    /**
-     * Don't use directly, instead call {@link #getGitHubUrl(String, ShipkitConfiguration)}.
-     * This way, we use lenient configuration and don't fail when auth token is not provided in shipkit configuration.
-     */
-    @ExposedForTesting
-    static String _getGitHubUrl(String ghUser, String ghRepo, String writeToken) {
-        if (writeToken != null) {
-            return MessageFormat.format("https://{0}:{1}@github.com/{2}.git", ghUser, writeToken, ghRepo);
-        } else {
-            return MessageFormat.format("https://github.com/{0}.git", ghRepo);
-        }
     }
 
     //TODO it's not longer only auth... let's find a better name
